+++ doc/CHANGES.rst
@@ -16,6 +16,8 @@ Bugs Fixed
 Features Added
 ++++++++++++++
 
+- LP #193122: New method getVirtualRoot added to the Request class
+
 - Added forward-compatibility shims for some frequently used modules moved in
   Zope 2.13.